+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Broadcast max steps as well when using single_host_load_and_broadcast.
+
 ## [0.11.23] - 2025-08-18
 
 ### Added